requirements.txt
pandas>=3.0.0
requests>=2.32.5
python-dotenv>=1.2.1
sqlalchemy>=2.0.46
pyarrow>=23.0.0
psycopg[binary]>=3.1